refactor: [EL-5010] refactor code to follow fsd rule
