+++ src/[fsd]/features/chat/lib/constants/index.js
@@ -1,2 +1,1 @@
 export * as ParticipantConstants from './participant.constants';
-export * as VoiceConstants from './voice.constants';
