+++ src/[fsd]/features/chat/lib/hooks/index.js
@@ -17,4 +17,3 @@ export {
   useNewStartConversationInputKeyDownHandler,
 } from './useInputKeyDownHandler.hooks';
 export { useDeleteMessageAlert } from './useDeleteMessageAlert.hooks';
-export { useVoiceConfig } from './useVoiceConfig.hooks';
