+++ src/[fsd]/features/chat/ui/index.js
@@ -5,7 +5,6 @@ export { ChatContinue } from './chat-continue';
 export { ChatHitlActions } from './chat-hitl-actions';
 export { SlashSuggestionList } from './slash-suggestion-list';
 export { VoiceMiniPlayer } from './voice-mini-player';
-export { VoiceConfigControls, VoiceConfigDialog } from './voice-config';
+export { VoiceConfigControls, VoiceConfigDialog, VoicePersonalizationSection } from '../voice-config';
 export { VoiceControlButton } from './voice-control-button';
-export { VoicePersonalizationSection } from './voice-personalization';
 export { ChatBox } from './chat-box';
