--- src/[fsd]/features/chat/ui/voice-config/index.js
@@ -1,2 +1,0 @@
-export { VoiceConfigControls } from './VoiceConfigControls';
-export { VoiceConfigDialog } from './VoiceConfigDialog';
