--- src/[fsd]/features/chat/ui/voice-personalization/index.js
@@ -1,1 +1,0 @@
-export { VoicePersonalizationSection } from './VoicePersonalizationSection';
