+++ src/[fsd]/features/chat/voice-config/index.js
@@ -1,0 +1,4 @@
+export { VoiceConfigControls } from './ui/VoiceConfigControls';
+export { VoiceConfigDialog } from './ui/VoiceConfigDialog';
+export { VoicePersonalizationSection } from './ui/VoicePersonalizationSection';
+export { useVoiceConfig } from './lib/hooks/useVoiceConfig.hooks';
