+++ src/[fsd]/features/chat/voice-config/ui/VoiceConfigControls.jsx
@@ -2,10 +2,11 @@ import { memo, useCallback, useMemo } from 'react';
 
 import { Box, Slider, Typography } from '@mui/material';
 
-import { VoiceConstants } from '@/[fsd]/features/chat/lib/constants';
 import { useTextToSpeech } from '@/[fsd]/features/chat/lib/hooks';
 import { Button } from '@/[fsd]/shared/ui';
 import SingleSelect from '@/[fsd]/shared/ui/select/SingleSelect';
+
+import * as VoiceConstants from '../constants/voice.constants';
 
 const { VOICE_PREVIEW_TEXT, VOICE_SPEED_MARKS, VOICE_VOLUME_MARKS } = VoiceConstants;
